requirements.txt
openai
crewai
python-telegram-bot
discord.py
fastapi
uvicorn
pydantic
pandas
pytest
requests
python-dotenv
pdoc
supervisor
termcolor
bandit
black
flake8
mypy
autoflake
isort
apscheduler
watchdog
celery
llama-cpp-python
huggingface_hub
playwright
